+++ instrumentation/spring/spring-scheduling-3.1/src/main/java/io/opentelemetry/instrumentation/auto/spring/scheduling/SpringSchedulingRunnableWrapper.java
@@ -1,0 +1,60 @@
+/*
+ * Copyright The OpenTelemetry Authors
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package io.opentelemetry.instrumentation.auto.spring.scheduling;
+
+import static io.opentelemetry.instrumentation.auto.spring.scheduling.SpringSchedulingDecorator.TRACER;
+import static io.opentelemetry.trace.TracingContextUtils.currentContextWith;
+
+import io.opentelemetry.context.Scope;
+import io.opentelemetry.trace.Span;
+
+public class SpringSchedulingRunnableWrapper implements Runnable {
+  private final Runnable runnable;
+
+  private SpringSchedulingRunnableWrapper(Runnable runnable) {
+    this.runnable = runnable;
+  }
+
+  @Override
+  public void run() {
+    if (runnable == null) {
+      return;
+    }
+    Span span =
+        TRACER.spanBuilder(SpringSchedulingDecorator.DECORATE.spanNameOnRun(runnable)).startSpan();
+    SpringSchedulingDecorator.DECORATE.afterStart(span);
+
+    try (Scope scope = currentContextWith(span)) {
+      runnable.run();
+    } catch (Throwable throwable) {
+      SpringSchedulingDecorator.DECORATE.onError(span, throwable);
+      throw throwable;
+    } finally {
+      SpringSchedulingDecorator.DECORATE.beforeFinish(span);
+      span.end();
+    }
+  }
+
+  public static Runnable wrapIfNeeded(Runnable task) {
+    // We wrap only lambdas' anonymous classes and if given object has not already been wrapped.
+    // Anonymous classes have '/' in class name which is not allowed in 'normal' classes.
+    if (task instanceof SpringSchedulingRunnableWrapper) {
+      return task;
+    }
+    return new SpringSchedulingRunnableWrapper(task);
+  }
+}
